Moved DarkThemePreference and DarkThemeProvider into AppTheme
