--- Flutter/lib/DarkThemeProvider.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'DarkThemePref.dart';
-
-class DarkThemeProvider with ChangeNotifier {
-  DarkThemePreference darkThemePreference = DarkThemePreference();
-  static bool _darkTheme = false;
-
-  bool get darkTheme => _darkTheme;
-
-  set darkTheme(bool value) {
-    _darkTheme = value;
-    darkThemePreference.setDarkTheme(value);
-    notifyListeners();
-  }
-}
